1/Add Position Class-2/Set property of each type of enemy in Program.Config-3/Add constructor for all file in package Enemy-4/Add some method for Enemy.Enemy : doDamage(Player), doDestroy, beAttacked, move(int, int), onUpdate

+++ src/Program/Player.java
@@ -12,7 +12,7 @@ public class Player {
         if (health <= 0) lose = true;
     }
 
-    public void beAttacked(final int damage){
+    public void beAttacked(final double damage){
         this.health -= damage;
     }
